Added check to make sure the repo is accesible on Github Incremented version to 1.2.9

+++ script.sh
@@ -2,7 +2,7 @@
 set -e
 
 #Script version
-VERSION="1.2.8"
+VERSION="1.2.9"
 
 #Color to the people
 RED='\x1B[0;31m'
@@ -362,11 +362,17 @@ if [ "$DBQUERY" -eq "1" ]; then
   
   echo -e "${GREEN}---> Fetching the latest version of the sripts...${NC}"
   echo -e
-  #Now let's fetch the latest version of the scripts
-  cd $SCRIPTPATH
-  git reset --hard HEAD
-  git pull
   
+  #First let's check if the repo is accesible
+  REPO_STATUS=$(curl -I "https://github.com/ElrondNetwork/elrond-go-scripts-v2" 2>&1 | awk '/HTTP\// {print $2}')
+  if [ "$REPO_STATUS" -eq "200" ]; then
+                                #Now let's fetch the latest version of the scripts
+                                echo -e "${GREEN}---> elrond-go-scripts-v2 is reachable ! Pulling latest version...${NC}"
+                                cd $SCRIPTPATH
+                                git reset --hard HEAD
+                                git pull
+                      else echo -e "${RED}---> elrond-go-scripts-v2 on Github not reachable !${NC}"
+              fi
   #Restore configs after repo pull
   echo -e "${GREEN}---> Restoring your config files${NC}"
   echo -e
